% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_annotation_utils.R
\name{calculate_smart_text_size}
\alias{calculate_smart_text_size}
\title{Smart Text Size Calculator}
\usage{
calculate_smart_text_size(n_items, base_size = 10, min_size = 8, max_size = 14)
}
\arguments{
\item{n_items}{Number of items to display}

\item{base_size}{Base text size}

\item{min_size}{Minimum text size}

\item{max_size}{Maximum text size}
}
\value{
Calculated text size
}
\description{
Smart Text Size Calculator
}
