% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_heatmap.R
\name{create_dendrogram}
\alias{create_dendrogram}
\title{Create dendrogram plot from hierarchical clustering}
\usage{
create_dendrogram(
  hclust_obj,
  dendro_line_size = 0.5,
  dendro_labels = FALSE,
  horizontal = FALSE
)
}
\arguments{
\item{hclust_obj}{An hclust object from hierarchical clustering}

\item{dendro_line_size}{Line width for dendrogram branches}

\item{dendro_labels}{Whether to show labels on dendrogram}

\item{horizontal}{Whether to create horizontal dendrogram}
}
\value{
A ggplot dendrogram
}
\description{
Create dendrogram plot from hierarchical clustering
}
\keyword{internal}
