% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_annotation_utils.R
\name{create_legend_theme}
\alias{create_legend_theme}
\title{Create Enhanced Legend Theme}
\usage{
create_legend_theme(
  position = "top",
  direction = "horizontal",
  title = NULL,
  title_size = 12,
  text_size = 10,
  key_size = 0.8,
  key_width = NULL,
  key_height = NULL,
  ncol = NULL,
  nrow = NULL,
  box_just = "center",
  margin = ggplot2::margin(0, 0, 0, 0)
)
}
\arguments{
\item{position}{Legend position ("top", "bottom", "left", "right", "none")}

\item{direction}{Legend direction ("horizontal", "vertical")}

\item{title}{Legend title}

\item{title_size}{Title font size}

\item{text_size}{Text font size}

\item{key_size}{Key size in cm}

\item{key_width}{Key width}

\item{key_height}{Key height}

\item{ncol}{Number of columns}

\item{nrow}{Number of rows}

\item{box_just}{Legend box justification}

\item{margin}{Legend margin}
}
\value{
ggplot2 theme elements
}
\description{
Create Enhanced Legend Theme
}
