% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_annotation_utils.R
\name{create_pathway_class_theme}
\alias{create_pathway_class_theme}
\title{Create Pathway Class Annotation Theme}
\usage{
create_pathway_class_theme(
  text_size = "auto",
  text_color = "black",
  text_face = "bold",
  text_family = "sans",
  text_angle = 0,
  text_hjust = 0.5,
  text_vjust = 0.5,
  bg_color = NULL,
  bg_alpha = 0.2,
  position = "left"
)
}
\arguments{
\item{text_size}{Text size}

\item{text_color}{Text color}

\item{text_face}{Text face ("plain", "bold", "italic")}

\item{text_family}{Text family}

\item{text_angle}{Text angle in degrees}

\item{text_hjust}{Horizontal justification (0-1)}

\item{text_vjust}{Vertical justification (0-1)}

\item{bg_color}{Background color}

\item{bg_alpha}{Background alpha}

\item{position}{Annotation position ("left", "right", "none")}
}
\value{
List of annotation styling parameters
}
\description{
Create Pathway Class Annotation Theme
}
