% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_gsea.R
\name{.as_color_vector}
\alias{.as_color_vector}
\title{Internal: extract a normalized color vector}
\usage{
.as_color_vector(scale)
}
\description{
Internal: extract a normalized color vector
}
\keyword{internal}
