% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_gsea.R
\name{.build_continuous_scale}
\alias{.build_continuous_scale}
\title{Internal: build a continuous ggplot2 scale layer from colors or ggplot2 scale}
\usage{
.build_continuous_scale(
  aes = c("fill", "color"),
  scale = NULL,
  diverging = FALSE,
  midpoint = NULL,
  name = NULL
)
}
\description{
Internal: build a continuous ggplot2 scale layer from colors or ggplot2 scale
}
\keyword{internal}
