% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_gsea.R
\name{.build_discrete_fill_for_direction}
\alias{.build_discrete_fill_for_direction}
\title{Internal: build a discrete fill scale for barplot direction}
\usage{
.build_discrete_fill_for_direction(scale = NULL)
}
\description{
Internal: build a discrete fill scale for barplot direction
}
\keyword{internal}
