% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_gsea.R
\name{.build_heatmap_col_fun}
\alias{.build_heatmap_col_fun}
\title{Internal: build a circlize colorRamp2 function for ComplexHeatmap from user scale}
\usage{
.build_heatmap_col_fun(scale = NULL)
}
\description{
Internal: build a circlize colorRamp2 function for ComplexHeatmap from user scale
}
\keyword{internal}
