% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_gsea.R
\name{.is_ggplot_scale}
\alias{.is_ggplot_scale}
\title{Internal: detect if an object is a ggplot2 Scale}
\usage{
.is_ggplot_scale(x)
}
\description{
Internal: detect if an object is a ggplot2 Scale
}
\keyword{internal}
