% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_themes.R
\name{get_available_themes}
\alias{get_available_themes}
\title{Get Available Color Themes}
\usage{
get_available_themes()
}
\value{
A character vector of available theme names
}
\description{
Get Available Color Themes
}
