% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_themes.R
\name{get_color_theme}
\alias{get_color_theme}
\title{Get Color Theme}
\usage{
get_color_theme(theme_name = "default", n_colors = 8)
}
\arguments{
\item{theme_name}{Character string specifying the theme name}

\item{n_colors}{Positive integer specifying the number of colors needed}
}
\value{
A list containing theme colors and settings
}
\description{
Get Color Theme
}
