% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ko_abundance.R
\docType{data}
\name{ko_abundance}
\alias{ko_abundance}
\title{KO Abundance Dataset}
\format{
A data frame where rownames are KO groups and column names include #NAME and individual sample names, such as:
"#NAME", "SRR11393730", "SRR11393731", "SRR11393732", "SRR11393733", "SRR11393734", "SRR11393735", "SRR11393736",
"SRR11393737", "SRR11393738", "SRR11393739", "SRR11393740", "SRR11393741", "SRR11393742", "SRR11393743",
"SRR11393744", "SRR11393745", "SRR11393746", "SRR11393747", "SRR11393748", "SRR11393749", "SRR11393750",
"SRR11393751", "SRR11393752", "SRR11393753", "SRR11393754", "SRR11393755", "SRR11393756", "SRR11393757",
"SRR11393758", "SRR11393759", "SRR11393760", "SRR11393761", "SRR11393762", "SRR11393763", "SRR11393764",
"SRR11393765", "SRR11393766", "SRR11393767", "SRR11393768", "SRR11393769", "SRR11393770", "SRR11393771",
"SRR11393772", "SRR11393773", "SRR11393774", "SRR11393775", "SRR11393776", "SRR11393777", "SRR11393778", "SRR11393779"
}
\source{
From ggpicrust2 package demonstration.
}
\usage{
ko_abundance
}
\description{
This is a demonstration dataset from the ggpicrust2 package, representing the output of PICRUSt2.
Each row represents a KO (KEGG Orthology) group, and each column corresponds to a sample.
}
\references{
Douglas GM, Maffei VJ, Zaneveld J, Yurgel SN, Brown JR, Taylor CM, Huttenhower C, Langille MGI. PICRUSt2 for prediction of metagenome functions. Nat Biotechnol. 2020.
}
\keyword{datasets}
