% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_themes.R
\name{smart_color_selection}
\alias{smart_color_selection}
\title{Smart Color Selection}
\usage{
smart_color_selection(
  n_groups,
  has_pathway_class = FALSE,
  data_type = "abundance",
  accessibility_mode = FALSE
)
}
\arguments{
\item{n_groups}{Number of groups in the data}

\item{has_pathway_class}{Whether pathway class information is available}

\item{data_type}{Type of data ("abundance", "pvalue", "foldchange")}

\item{accessibility_mode}{Whether to require accessibility-friendly colors.
When TRUE, this constraint takes precedence over data-type heuristics.}
}
\value{
A list with suggested theme and colors
}
\description{
Intelligently selects colors based on data characteristics
}
