% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_daa.R
\name{validate_abundance_data}
\alias{validate_abundance_data}
\title{Validate abundance data for PICRUSt compatibility}
\usage{
validate_abundance_data(abundance_mat, daa_method)
}
\arguments{
\item{abundance_mat}{Abundance matrix}

\item{daa_method}{DAA method being used}
}
\description{
Validate abundance data for PICRUSt compatibility
}
\keyword{internal}
