% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_validation.R
\name{validate_gene_set_quality}
\alias{validate_gene_set_quality}
\title{Validate gene set quality metrics}
\usage{
validate_gene_set_quality(gene_sets, pathway_type)
}
\arguments{
\item{gene_sets}{List of gene sets}

\item{pathway_type}{Character string specifying pathway type}
}
\description{
Validate gene set quality metrics
}
\keyword{internal}
