% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_gsea.R
\name{.as_color_vector}
\alias{.as_color_vector}
\title{Internal: coerce user 'scale' input to a vector of colors (or NULL)
Accepts character vector or function(n)->colors. Returns character vector or NULL.}
\usage{
.as_color_vector(scale)
}
\description{
Internal: coerce user 'scale' input to a vector of colors (or NULL)
Accepts character vector or function(n)->colors. Returns character vector or NULL.
}
\keyword{internal}
