% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpicrust2.R
\name{ggpicrust2}
\alias{ggpicrust2}
\title{Title}
\usage{
ggpicrust2(
  file,
  metadata,
  group,
  pathway,
  daa_method = "ALDEx2",
  ko_to_kegg = FALSE,
  p.adjust = "BH",
  order = "group",
  p_values_bar = TRUE,
  x_lab = "pathway_name",
  select = NULL,
  reference = NULL,
  colors = NULL
)
}
\arguments{
\item{colors}{}
}
\description{
Title
}
