% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ko2kegg_abundance.R
\name{ko2kegg_abundance}
\alias{ko2kegg_abundance}
\title{Convert KO abundance in picrust2 export files to KEGG pathway abundance}
\usage{
ko2kegg_abundance(file)
}
\arguments{
\item{file}{A character, the address of the file containing KO abundance in picrust2 export format}
}
\description{
Takes a file containing KO abundance in picrust2 export format and converts it to KEGG pathway abundance.
The file should be in .tsv, .txt or .csv format.
}
\examples{
\dontrun{
library(ggpicrust2)
# Prepare an example input file
input_file <- system.file("extdata", "example_ko_abundance.tsv", package = "ggpicrust2")

# Run ko2kegg_abundance function
kegg_abundance <- ko2kegg_abundance(file = input_file)
}
}
