% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ko2kegg_abundance.R
\name{ko2kegg_abundance}
\alias{ko2kegg_abundance}
\title{Title This function converts the KO abundance in picrust2 export files to KEGG pathway abundance}
\usage{
ko2kegg_abundance(file)
}
\arguments{
\item{file}{A character, address to store KO abundance in picrust2 export files}
}
\value{
A data frame, kegg_abundance
}
\description{
Title This function converts the KO abundance in picrust2 export files to KEGG pathway abundance
}
