% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_annotation.R
\name{pathway_annotation}
\alias{pathway_annotation}
\title{Pathway information annotation of "EC", "KO", "MetaCyc" pathway}
\usage{
pathway_annotation(
  file = NULL,
  pathway = NULL,
  daa_results_df = NULL,
  ko_to_kegg = FALSE
)
}
\arguments{
\item{file}{A character, address to store picrust2 export files}

\item{pathway}{A character, consisting of "EC", "KO", "MetaCyc"}

\item{daa_results_df}{A data frame, output of pathway_daa}

\item{ko_to_kegg}{A character, decide if convert ko abundance to kegg pathway abundance}
}
\description{
Pathway information annotation of "EC", "KO", "MetaCyc" pathway
}
