% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_daa.R
\name{pathway_daa}
\alias{pathway_daa}
\title{Title}
\usage{
pathway_daa(
  abundance,
  metadata,
  group,
  daa_method,
  select = NULL,
  p.adjust = "BH",
  reference = NULL
)
}
\arguments{
\item{reference}{}
}
\description{
Title
}
