% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_daa.R
\name{pathway_daa}
\alias{pathway_daa}
\title{Predictional functional patwhay differential abundance (DA)}
\usage{
pathway_daa(
  abundance,
  metadata,
  group,
  daa_method = "ALDEx2",
  select = NULL,
  p.adjust = "BH",
  reference = NULL
)
}
\arguments{
\item{abundance}{a data frame containing predicted functional pathway abundance}

\item{metadata}{a tibble containing samples information}

\item{group}{a character specifying the group name for differential abundance analysis}

\item{daa_method}{a character specifying the method for differential abundance analysis, default is "ALDEx2"}

\item{select}{a vector containing pathway names for analysis, if NULL all pathways are included, default is NULL}

\item{p.adjust}{a character specifying the method for p-value adjustment, default is "BH"}

\item{reference}{a character specifying the reference group level, required for several differential abundance analysis methods, default is NULL}
}
\value{
a data frame containing the differential abundance analysis results.
}
\description{
Predictional functional patwhay differential abundance (DA)
}
