% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_errorbar.R
\name{pathway_errorbar}
\alias{pathway_errorbar}
\title{The function pathway_errorbar() is used to visualize the predictional functional pathway differential abundance analysis results as error bar plots.}
\arguments{
\item{abundance}{a data frame where the row names represent pathways and the column names represent samples, with each element representing the relative abundance of the corresponding pathway in the corresponding sample.}

\item{daa_results_df}{a data frame of the results of the differential abundance analysis of the pathways, generated by the pathway_daa function.}

\item{Group}{a data frame or a vector that assigns each sample to a group, which is used to color the samples in the figure.}

\item{ko_to_kegg}{a logical parameter indicating whether to convert the pathway IDs from KEGG orthology (KO) to KEGG pathway IDs.}

\item{p_values_threshold}{a numeric parameter controlling the threshold for statistical significance of differential abundance.}

\item{order}{a parameter controlling the ordering of the rows in the figure, which can be "p_values" (order by p-values), "name" (order by pathway name), "group" (order by the group with the highest mean relative abundance), or "pathway_class" (order by the pathway category).}

\item{select}{a vector of pathway names to be included in the figure, which can be used to limit the number of pathways displayed.}

\item{p_value_bar}{a logical parameter indicating whether to display a bar showing the p-value threshold for significance.}

\item{colors}{a vector of colors to be used to represent the groups in the figure.}

\item{x_lab}{a character string to be used as the x-axis label in the figure, which defaults to "description" for KO IDs and "pathway_name" for KEGG pathway IDs.}
}
\value{
A figure
}
\description{
The function pathway_errorbar() is used to visualize the predictional functional pathway differential abundance analysis results as error bar plots.
}
