% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_errorbar.R
\name{pathway_errorbar}
\alias{pathway_errorbar}
\title{The function pathway_errorbar() is used to visualize the results of functional pathway differential abundance analysis as error bar plots.}
\arguments{
\item{abundance}{A data frame with row names representing pathways and column names representing samples. Each element represents the relative abundance of the corresponding pathway in the corresponding sample.}

\item{daa_results_df}{A data frame containing the results of the differential abundance analysis of the pathways, generated by the pathway_daa function.}

\item{Group}{A data frame or a vector that assigns each sample to a group. The groups are used to color the samples in the figure.}

\item{ko_to_kegg}{A logical parameter indicating whether to convert pathway IDs from KEGG orthology (KO) to KEGG pathway IDs. If TRUE, the conversion will be performed.}

\item{p_values_threshold}{A numeric parameter specifying the threshold for statistical significance of differential abundance. Pathways with p-values below this threshold will be considered significant.}

\item{order}{A parameter controlling the ordering of the rows in the figure. The options are: "p_values" (order by p-values), "name" (order by pathway name), "group" (order by the group with the highest mean relative abundance), or "pathway_class" (order by the pathway category).}

\item{select}{A vector of pathway names to be included in the figure. This can be used to limit the number of pathways displayed. If NULL, all pathways will be displayed.}

\item{p_value_bar}{A logical parameter indicating whether to display a bar showing the p-value threshold for significance. If TRUE, the bar will be displayed.}

\item{colors}{A vector of colors to be used to represent the groups in the figure. Each color corresponds to a group.}

\item{x_lab}{A character string to be used as the x-axis label in the figure. The default value is "description" for KO IDs and "pathway_name" for KEGG pathway IDs.}
}
\value{
A figure showing the error bar plot of the differential abundance analysis results for the functional pathways.
}
\description{
The function pathway_errorbar() is used to visualize the results of functional pathway differential abundance analysis as error bar plots.
}
