% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_heatmap.R
\name{pathway_heatmap}
\alias{pathway_heatmap}
\title{Create pathway heatmap}
\usage{
pathway_heatmap(abundance, metadata, group)
}
\arguments{
\item{abundance}{A data frame, predicted functional pathway abundance}

\item{metadata}{A tibble, consisting of samples information}

\item{group}{A character, group name}
}
\value{
A ggplot heatmap object
}
\description{
This function creates a heatmap of the predicted functional pathway abundance data. The function first makes the abundance data relative, then converts the abundance data to a long format and orders the samples based on the environment information. The heatmap is then created using the `ggplot2` library. The color palette, appearance and the color bar of the heatmap can be customized using the `scale_fill_gradientn`, `theme` and `guides` functions respectively.
}
