% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_gsea.R
\name{prepare_gene_sets}
\alias{prepare_gene_sets}
\title{Prepare gene sets for GSEA}
\usage{
prepare_gene_sets(pathway_type = "KEGG", organism = "ko")
}
\arguments{
\item{pathway_type}{A character string specifying the pathway type: "KEGG", "MetaCyc", or "GO"}

\item{organism}{A character string specifying the organism (only relevant for KEGG and GO)}
}
\value{
A list of pathway gene sets
}
\description{
Prepare gene sets for GSEA
}
\keyword{internal}
