% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_gsea.R
\name{run_fgsea}
\alias{run_fgsea}
\title{Run fgsea using the recommended fgseaMultilevel method}
\usage{
run_fgsea(ranked_list, gene_sets, min_size = 10, max_size = 500)
}
\arguments{
\item{ranked_list}{A named vector of ranking statistics}

\item{gene_sets}{A list of pathway gene sets}

\item{min_size}{An integer specifying the minimum gene set size}

\item{max_size}{An integer specifying the maximum gene set size}
}
\value{
A data frame of fgsea results
}
\description{
Run fgsea using the recommended fgseaMultilevel method
}
\keyword{internal}
